requirements.txt
streamlit>=1.30
opencv-python>=4.8
torch>=2.1
torchvision>=0.16
numpy>=1.24
ultralytics>=8.0
segmentation-models-pytorch>=0.3
Pillow>=10.0
pycocotools>=2.0
torchmetrics>=1.0
pandas>=2.0